Initial commit of EuclideanSpace framework, Algorithms, Constants, Point2, Point3, Vector2, and Vector3
Algorithms.swift has functions to compute
	- euclidean distance
	- dot product
	- convert from rad to deg
	- convert from deg to rad

Constants.swift is empty for now, but may be used to store protocols, and of course, constants.

Point2/Point3 consists of a [Double] field; they are coordinates on their respective Cartesian plane.
Vector2/Vector3 also consists of a [Double] field; they are vector components.

As of this commit, the package builds and can be added to an Xcode project.